test: move compiler tests to unit test directory
