+++ .github/workflows/main.yml
@@ -40,3 +40,22 @@ jobs:
     - name: isort
       run: isort --check-only .
       if: always()
+
+  unit-tests:
+    name: Unit Tests
+    runs-on: ubuntu-latest
+    steps:
+
+    - name: checkout
+      uses: actions/checkout@v1
+
+    - name: setup python
+      uses: actions/setup-python@v2
+      with:
+        python-version: "3.7"
+
+    - name: install dependencies
+      run: ./ci/install_deps.sh
+
+    - name: run tests
+      run: python -m pytest tests/unit
